// src/app/shared/util-model/enumeration/event-option.enum.ts
export enum EventOptionEnum {
    TICKETING = 'TICKETING',
    VEHICLE = 'VEHICLE',
    ACTIVITY = 'ACTIVITY',
    PHONE_COMMUNICATION = 'PHONE_COMMUNICATION',
    ACTIVITY_COMMUNICATION = 'ACTIVITY_COMMUNICATION',
    FIRE_RISK = 'FIRE_RISK',
    SMOKE_REPORT = 'SMOKE_REPORT',
    MOVEMENT_REPORT = 'MOVEMENT_REPORT',
}